requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
azure-ai-nspkg
azure-applicationinsights~=0.1.0
azure-batch~=4.1
azure-cognitiveservices-nspkg
azure-cognitiveservices-knowledge-nspkg
azure-cognitiveservices-language-nspkg
azure-cognitiveservices-search-nspkg
azure-cognitiveservices-vision-nspkg
azure-common~=1.1
azure-core<2.0.0,>=1.2.2
azure-cosmosdb-table~=1.0
azure-datalake-store~=0.0.18
azure-eventhub<6.0.0,>=5.0.0
azure-eventgrid~=1.1
azure-loganalytics~=0.1.0
azure-mgmt~=4.0
azure-mgmt-advisor~=1.0
azure-mgmt-applicationinsights~=0.1.1
azure-mgmt-authorization~=0.50.0
azure-mgmt-batch~=5.0
azure-mgmt-batchai~=2.0
azure-mgmt-billing~=0.2.0
azure-mgmt-cdn~=3.0
azure-mgmt-cognitiveservices~=3.0
azure-mgmt-commerce~=1.0
azure-mgmt-compute~=4.0
azure-mgmt-consumption~=2.0
azure-mgmt-containerinstance~=1.0
azure-mgmt-containerregistry~=2.1
azure-mgmt-containerservice~=4.2
azure-mgmt-cosmosdb~=0.4.1
azure-mgmt-datafactory~=0.6.0
azure-mgmt-datalake-analytics~=0.6.0
azure-mgmt-datalake-nspkg
azure-mgmt-datalake-store~=0.5.0
azure-mgmt-datamigration~=1.0
azure-mgmt-devspaces~=0.1.0
azure-mgmt-devtestlabs~=2.2
azure-mgmt-dns~=2.0
azure-mgmt-eventgrid~=1.0
azure-mgmt-eventhub~=2.1
azure-mgmt-hanaonazure~=0.1.1
azure-mgmt-iotcentral~=0.1.0
azure-mgmt-iothub~=0.5.0
azure-mgmt-iothubprovisioningservices~=0.2.0
azure-mgmt-keyvault~=1.0
azure-mgmt-loganalytics~=0.2.0
azure-mgmt-logic~=3.0
azure-mgmt-machinelearningcompute~=0.4.1
azure-mgmt-managementgroups~=0.1.0
azure-mgmt-managementpartner~=0.1.0
azure-mgmt-maps~=0.1.0
azure-mgmt-marketplaceordering~=0.1.0
azure-mgmt-media~=1.0.0rc2
azure-mgmt-monitor~=0.5.2
azure-mgmt-msi~=0.2.0
azure-mgmt-network~=2.0
azure-mgmt-notificationhubs~=2.0
azure-mgmt-nspkg
azure-mgmt-policyinsights~=0.1.0
azure-mgmt-powerbiembedded~=2.0
azure-mgmt-rdbms~=1.2
azure-mgmt-recoveryservices~=0.3.0
azure-mgmt-recoveryservicesbackup~=0.3.0
azure-mgmt-redis~=5.0
azure-mgmt-relay~=0.1.0
azure-mgmt-reservations~=0.2.1
azure-mgmt-resource~=2.0
azure-mgmt-scheduler~=2.0
azure-mgmt-search~=2.0
azure-mgmt-servicebus~=0.5.1
azure-mgmt-servicefabric~=0.2.0
azure-mgmt-signalr~=0.1.0
azure-mgmt-sql~=0.9.1
azure-mgmt-storage~=2.0
azure-mgmt-subscription~=0.2.0
azure-mgmt-trafficmanager~=0.50.0
azure-mgmt-web~=0.35.0
